move: main and index files
