+++ static_report/src/App.tsx
@@ -1,15 +1,19 @@
 import { Suspense, lazy } from 'react';
 import { Switch, Route, Router, BaseLocationHook } from 'wouter';
 
-import { Main } from './components/Main';
+import { Main } from './components/shared/Main';
 import { Loading } from './components/shared/Loading';
 import { NotFound } from './components/shared/NotFound';
 import { useHashLocation } from './hooks/useHashLcocation';
 import { SingleReportList } from './components/SingleReport/SRList';
 import { ComparisonReportList } from './components/ComparisonReport/CRList';
 
-const SingleReport = lazy(() => import('./components/SingleReport'));
-const ComparisonReport = lazy(() => import('./components/ComparisonReport'));
+const SingleReport = lazy(
+  () => import('./components/SingleReport/SingleReport'),
+);
+const ComparisonReport = lazy(
+  () => import('./components/ComparisonReport/ComparisonReport'),
+);
 
 function AppSingle() {
   const { tables, datasource } = window.PIPERIDER_SINGLE_REPORT_DATA;
