+++ static_report/src/components/shared/Main.tsx
